+++ src/Contract/Client/Exception/FactorizeClientProviderNotFoundException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Heptacom\AdminOpenAuth\Contract\Client\Exception;
-
-final class FactorizeClientProviderNotFoundException extends FactorizeClientException
-{
-}